LayoutTests/imported/w3c/web-platform-tests/workers/SharedWorker_dataUrl-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected 2 but got 1


Harness Error (FAIL), message = Error: assert_equals: expected 2 but got 1

PASS Data URL not shared by cross-origin SharedWorkers
TIMEOUT Data URLs shared by same-origin SharedWorkers Test timed out

Harness Error (FAIL), message = Error: assert_equals: expected 2 but got 1

PASS Data URL not shared by cross-origin SharedWorkers
TIMEOUT Data URLs shared by same-origin SharedWorkers Test timed out
